Refactor : 그룹이름, 메뉴바 통합
